+++ tests/lib/managers/files-test.js
@@ -1074,6 +1074,30 @@ describe('Files', function() {
 			files.uploadFile(PARENT_FOLDER_ID, FILENAME, CONTENT);
 		});
 
+		it('should pass optional parameters in attributes when specified', function() {
+
+			var options = {
+				content_modified_at: '2017-11-18T11:18:00-0800',
+				content_created_at: '1988-11-18T11:18:00-0800'
+			};
+			var expectedFormData = {
+				attributes: JSON.stringify({
+					name: FILENAME,
+					parent: { id: PARENT_FOLDER_ID },
+					content_modified_at: options.content_modified_at,
+					content_created_at: options.content_created_at
+				}),
+				content: {
+					value: CONTENT,
+					options: { filename: 'unused' }
+				}
+			};
+
+			sandbox.stub(boxClientFake, 'wrapWithDefaultHandler').returnsArg(0);
+			sandbox.mock(boxClientFake).expects('upload').withArgs('/files/content', null, expectedFormData);
+			files.uploadFile(PARENT_FOLDER_ID, FILENAME, CONTENT, options);
+		});
+
 		it('should wrap with default handler when called', function() {
 
 			sandbox.stub(boxClientFake, 'upload').returns(Promise.resolve());
@@ -1119,6 +1143,25 @@ describe('Files', function() {
 			sandbox.stub(boxClientFake, 'wrapWithDefaultHandler').returnsArg(0);
 			sandbox.mock(boxClientFake).expects('upload').withArgs('/files/1234/content', null, expectedFormData);
 			files.uploadNewFileVersion(FILE_ID, CONTENT);
+		});
+
+		it('should pass optional parameters in attributes when specified', function() {
+
+			var options = {
+				name: 'New filename.txt',
+				content_modified_at: '2017-12-17T12:34:56-0800'
+			};
+			var expectedFormData = {
+				attributes: JSON.stringify(options),
+				content: {
+					value: CONTENT,
+					options: { filename: 'unused' }
+				}
+			};
+
+			sandbox.stub(boxClientFake, 'wrapWithDefaultHandler').returnsArg(0);
+			sandbox.mock(boxClientFake).expects('upload').withArgs('/files/1234/content', null, expectedFormData);
+			files.uploadNewFileVersion(FILE_ID, CONTENT, options);
 		});
 
 		it('should wrap with default handler when called', function() {